verification: split the trusted-path POC into per-fix modules
One 200-line lib.rs became one small module per fix (frame_alloc,
input_authority, reply_correlation, register_auth, framebuffer) with the tests
moved to the tests directory, so lib.rs only declares the modules and no file is
over the size limit. The eight checks are unchanged.
